+++ python/featomic/tests/clebsch_gordan/equivariant_power_spectrum_by_pair.py
@@ -136,11 +136,9 @@ def test_sample_selection() -> None:
         SphericalExpansionByPair(**SPHEX_HYPERS_SMALL),
     )
 
-    label_1st = mts.Labels(["system", "atom"], np.array([[0, 0]], dtype=np.int32))
+    label_1st = mts.Labels(["system", "atom"], np.array([[0, 0]]))
 
-    label_2nd = mts.Labels(
-        ["system", "atom"], np.array([[0, 1], [0, 2]], dtype=np.int32)
-    )
+    label_2nd = mts.Labels(["system", "atom"], np.array([[0, 1], [0, 2]]))
 
     powspec_1 = powspec_by_pair_calc.compute(
         frame, neighbors_to_properties=True, selected_samples=label_1st
